PROTO version bump, corrections
serilog inverts debug / verbose. The internal code now accounts for that
because its too painful to visual switch it while reading the log file.

the current version bump was due to the experimentation with
publishing to powershellgallery.org.

The build.ps1 file uses devenv.exe to build the project
because the csproj basically wraps project.json which
has issues using vanilla msbuild.